refactor(facade): the door's two constructions get their own files
Surface keeps what it is — install, and the RESERVED words — and hands
the two module constructions to surface/chapter.rb (a chapter's module,
with its declaration hook for names the door does not carry) and
surface/aggregate_door.rb (one aggregate's creating verbs, CRUD
delegation, and the hexagon binding collector).
